wire up google maps + socrata endpoint example
This is a nice little snippet that Jillian found here:

   http://jsfiddle.net/chrismetcalf/8m2Cs/

It illustrates how to position markers on a google map
based upon lat/lng data from another data source.

The two data sources we're currently exploring are also available
as an Socrata Open Data API (SODA) so this is higly relevant.