fix issue with 0-height child map div
We notice that unless we add an explicit height to the .map class in
the css, it ends up with 0-height and thus does not appear on the
screen.

Wondering if this has to do with box-model?  I would have expected
the 480 height to be inherited by the child from the parent map-container
div.  Need to explore this more.