make a separate function, loadData, for fetching data source data such as:
	* crime data
	* traffic data
	* school district data

Before this was munged into the same code that pulled in the
google map.  Logically these two activities should be separate
since we may want filter crime data, for example, by
severity.

Also create a simple helper function for displaying the incoming
raw JSON data to a scrollable div at the bottom of the browser
for easier debugging.