wow, the location of the 2016 Austin traffic fatalities endpoint changed today
It used to be here:

   https://data.austintexas.gov/resource/s88d-6sad.json

According to:

   https://dev.socrata.com/foundry/data.austintexas.gov/vn6k-4eq5

The endpoint was updated at about 1pm today and the name got changed
in the process apparently.

   Dataset Identifier: vn6k-4eq5
   Total Rows: 60
   Source Domain: data.austintexas.gov
 >>Created: 11/21/2016, 1:09:30 PM
 >>Last Updated: 11/21/2016, 1:20:39 PM
   Category: Public Safety
   Owner: Ron MacKay

If this happens frequently, it could impact the usability of the
app.  Should probably open up an issue to track this problem
and find out how to manage this in an automated way.