handle edgecases better with regard to empty address processing
We have a few model methods that mutate street addresses.
It seems we're prone, at the moment, to hitting web api limits
in fetching traffic incident data.

This can result in a lot of 'undefined' addresses getting pumped
through our pipeline.  When this happens, don't try to slap a
city and state on the undefined address, else the caller will think
there is something legit to work with.  And upstream logic that
would otherwise filter out the undefined address fails.