fallback to dataSource that /doesn't/ require geocoding since that is hard
Last night, we brutally learned that it's one thing to geocode a single address.
It's entirely different to geocode 60 in a block with Google's geocode api.

After about the first 10 addresses, we get hit with:

	OVER_QUERY_LIMIT

and the other 50 addresses we want to fetch latitude and longitude for
get ignored.  This means we're seeing < 20% of the traffic incident data
we expect.

This will likely require a much more sophisticated approach with block
encoding, persisting to a db, and driving the app off of our own db.

But we don't really have time to do that before demo day, so we're falling
back to a data source that does have lat/lng already in the data set:

	trafficFatalities2016: {
		description: "2016 Austin Traffic Fatalities",
		queryUrl: "https://data.austintexas.gov/resource/vn6k-4eq5.json",
		apiKeyName: "",
		apiKey: ""
	}