make it easier to configure the app for places besides Austin
The application is largely model driven.  However we were still hardcoding
the jumbotron background image and text.  This commit makes both of these
attributes model properties that may be easily localized for new places
we add to the model.

See: model.js

var model = {
   places: {
      austin: {
        >appName: "Austin Aware",
        >backgroundUrl: "http://www.pictureofAustin.com",
        >backgroundImagePosition: "top center",
         ..
      }
      ..
   }
}

This /does/ make the model aware of the view, per-se, and maybe violate
some stricture of MVC.  But it's in the service of easily reconfiguring
the entire app, just by making a call to:

   model.setPlace("Timbuktu");

and then having the background change accordingly along with the
data sources that are relevant for this new place.