not all the records coming back from 2015 traffic data have the same structure
Consequently, we see this error pop up in the console inspector when
loadData is called:

   Uncaught TypeError: Cannot read property 'coordinates' of undefined(..)

Normally, the data coming back from the endpoint for 2015 Austin Traffic Fatalities
is an array of records that include details, such as latitude and longitude,
for accidents.

However, at the end of the dataset, there are two atypical records that
reflect how many accidents in total happened for the year and which
previous entries were ruled suicide instead of accident.

This fix adds some guard logic for this condition in the entry iterator.