push 'hover-text' for a given map marker into the model
Stuff happens in time and space.  It turns into a marker on a map.
We'd like to know something about what that marker represents.  You
know, the details.

But those details are often relegated to different places in the
inbound json records.  So we create custom 'detail-extractors'
for each data source within the model.

This allows us to call a high-level generic function like:

   var detailString = model.getMarkerTitle(jsonEntry, "trafficFatalities2015");

and it will drill down to the data source and figure out how
to construct the marker hover-over text for the 2015 fatalities data source.