reset zoom button should also re-center the map too
If user has zoomed in and/or panned the map, the reset
button should normalize both attributes to the model defaults.

It makes sense, then, to relabel the button as 'Reset Map' instead
of 'Reset Zoom'

I'm also taking the opportunity to tweak the styling a tad so the
bottom of the reset button lines up with the bottom of the map
in wide-screen mode.  Also making the button text white to
fall in line with other button text styling.

I've come around to DB's notion of contrasting the reset button
background color with the other data-selection buttons.  I'm
color picking a dark blue pulled from the skyline image in the
masthead.  This seems to work well with white text for contrast.