the deployed heroku image did not pick up the endpoint change; be more persuasive
This is very odd.  I think I'm hitting an heroku deployment bug.
For some reason, heroku deployed a down-level snapshot of the
repository that did /not/ include commit: 1fe337c

   update the endpoint for 2016 data
   These endpoints change on a monthly basis for the current year.

Hoping some whitespace tweak will force heroku to pick up a
new version of the file with the new endpoint.